Refactor websocket handlers into modular router
