+++ lib/ws-handlers/scenario-handlers.js
@@ -1,0 +1,9 @@
+export default function registerScenarioHandlers({ scenarioEngine, handlers }) {
+  scenarioEngine.on('scenarioStarted', (data) => {
+    handlers.broadcastScenarioEvent('scenario-started', data);
+  });
+
+  scenarioEngine.on('scenarioCompleted', (data) => {
+    handlers.broadcastScenarioEvent('scenario-completed', data);
+  });
+}
